+++ packages/nextclaw-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextclaw/remote
 
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @nextclaw/server@0.10.41
+
 ## 0.1.34
 
 ### Patch Changes